+++ internal_filesystem/lib/mpos/board/linux.py
@@ -89,13 +89,13 @@ except Exception as e:
 
 
 # Simulated battery voltage ADC measuring
-import mpos.battery_voltage
+from mpos import BatteryManager
 
 def adc_to_voltage(adc_value):
     """Convert simulated ADC value to voltage."""
     return adc_value * (3.3 / 4095) * 2
 
-mpos.battery_voltage.init_adc(999, adc_to_voltage)
+BatteryManager.init_adc(999, adc_to_voltage)
 
 # === AUDIO HARDWARE ===
 from mpos import AudioFlinger